refactor: move PrefetchProvider to components/ui, separate from QueryClientProvider
QueryClientProvider stays in entry-client.tsx at app root.
PrefetchProvider moved to src/components/ui/ to match project structure.
