+++ packages/chronicle/src/server/entry-client.tsx
@@ -5,7 +5,7 @@ import { BrowserRouter } from 'react-router';
 import { ReactRouterProvider } from 'fumadocs-core/framework/react-router';
 import { QueryClientProvider } from '@tanstack/react-query';
 import { mdxComponents } from '@/components/mdx';
-import { PrefetchProvider } from '@/components/PrefetchProvider';
+import { PrefetchProvider } from '@/components/ui/PrefetchProvider';
 import { getApiConfigsForVersion } from '@/lib/config';
 import { PageProvider } from '@/lib/page-context';
 import { queryClient } from '@/lib/preload';
